Resolve conflicts in Tests/CTestTestMemcheck/CMakeLists.txt by combining
changes from both sides.

+++ Source/kwsys/testSystemInformation.cxx
@@ -88,9 +88,9 @@ int testSystemInformation(int, char*[])
   printMethod3(info, GetHostMemoryUsed(), "KiB");
   printMethod3(info, GetProcMemoryUsed(), "KiB");
 
-  for (int i = 0; i <= 31; i++)
+  for (long int i = 0; i <= 31; i++)
     {
-    if (info.DoesCPUSupportFeature(1 << i))
+    if (info.DoesCPUSupportFeature(static_cast<long int>(1) << i))
       {
       kwsys_ios::cout << "CPU feature " << i << "\n";
       }
